Lab2
I chose 3 combinations. I first went with the default price versus sqft just so there was a high R^2 value to compare against. Then I tried combining Beds and bath versus price to see if there was a correlation there. It was an R^2 between 0.3 and 0.4, so not abhorrent, but not good either. For my final graph, I was going to do a combination of all three bed, bath, and property sqft. I stopped doing this after realizing, no matter what coefficient I put on it, having beds included seemed to reduce the R^2 value. So I ended up using the sum of 500*beds and property sqft. This resulted in a R^2 of nearly 0.6, so I was pretty satisfied